Move LSPS2 client logic into `liquidity/client/lsps2.rs`
