+++ src/liquidity/mod.rs
@@ -14,61 +14,41 @@ use std::ops::Deref;
 use std::sync::{Arc, Mutex, RwLock, Weak};
 use std::time::Duration;
 
-use bitcoin::secp256k1::{PublicKey, Secp256k1};
+use bitcoin::secp256k1::PublicKey;
 use bitcoin::Transaction;
 use chrono::Utc;
 use lightning::events::HTLCHandlingFailureType;
-use lightning::ln::channelmanager::{InterceptId, MIN_FINAL_CLTV_EXPIRY_DELTA};
+use lightning::ln::channelmanager::InterceptId;
 use lightning::ln::msgs::SocketAddress;
 use lightning::ln::types::ChannelId;
-use lightning::routing::router::{RouteHint, RouteHintHop};
 use lightning::sign::EntropySource;
-use lightning::util::ser::Writeable;
-use lightning_invoice::{Bolt11Invoice, Bolt11InvoiceDescription, InvoiceBuilder, RoutingFees};
 use lightning_liquidity::events::LiquidityEvent;
-use lightning_liquidity::lsps0::ser::{LSPSDateTime, LSPSRequestId};
+use lightning_liquidity::lsps0::ser::LSPSDateTime;
 use lightning_liquidity::lsps1::client::LSPS1ClientConfig as LdkLSPS1ClientConfig;
 use lightning_liquidity::lsps2::client::LSPS2ClientConfig as LdkLSPS2ClientConfig;
 use lightning_liquidity::lsps2::event::{LSPS2ClientEvent, LSPS2ServiceEvent};
-use lightning_liquidity::lsps2::msgs::{LSPS2OpeningFeeParams, LSPS2RawOpeningFeeParams};
+use lightning_liquidity::lsps2::msgs::LSPS2RawOpeningFeeParams;
 use lightning_liquidity::lsps2::service::LSPS2ServiceConfig as LdkLSPS2ServiceConfig;
-use lightning_liquidity::lsps2::utils::compute_opening_fee;
 use lightning_liquidity::{LiquidityClientConfig, LiquidityServiceConfig};
 use lightning_types::payment::PaymentHash;
-use tokio::sync::oneshot;
 
 use crate::builder::BuildError;
-use crate::logger::{log_debug, log_error, log_info, LdkLogger};
-use crate::payment::store::LSPS2Parameters;
-use crate::payment::PaymentMetadata;
+use crate::logger::{log_error, LdkLogger};
 use crate::types::{
 	Broadcaster, ChannelManager, DynStore, KeysManager, LiquidityManager, PeerManager, Wallet,
 };
-use crate::{total_anchor_channels_reserve_sats, Config, Error};
+use crate::{total_anchor_channels_reserve_sats, Config};
 
 pub(crate) use client::lsps1::{LSPS1Client, LSPS1ClientConfig};
 pub use client::lsps1::{LSPS1Liquidity, LSPS1OrderStatus};
+pub(crate) use client::lsps2::{
+	LSPS2BuyResponse, LSPS2Client, LSPS2ClientConfig, LSPS2FeeResponse,
+};
 
 pub(crate) const LIQUIDITY_REQUEST_TIMEOUT_SECS: u64 = 5;
 
 const LSPS2_GETINFO_REQUEST_EXPIRY: Duration = Duration::from_secs(60 * 60 * 24);
 const LSPS2_CHANNEL_CLTV_EXPIRY_DELTA: u32 = 72;
-
-struct LSPS2Client {
-	lsp_node_id: PublicKey,
-	lsp_address: SocketAddress,
-	token: Option<String>,
-	ldk_client_config: LdkLSPS2ClientConfig,
-	pending_fee_requests: Mutex<HashMap<LSPSRequestId, oneshot::Sender<LSPS2FeeResponse>>>,
-	pending_buy_requests: Mutex<HashMap<LSPSRequestId, oneshot::Sender<LSPS2BuyResponse>>>,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct LSPS2ClientConfig {
-	pub node_id: PublicKey,
-	pub address: SocketAddress,
-	pub token: Option<String>,
-}
 
 struct LSPS2Service {
 	service_config: LSPS2ServiceConfig,
@@ -298,10 +278,6 @@ where
 
 	pub(crate) fn liquidity_manager(&self) -> Arc<LiquidityManager> {
 		Arc::clone(&self.liquidity_manager)
-	}
-
-	pub(crate) fn get_lsps2_lsp_details(&self) -> Option<(PublicKey, SocketAddress)> {
-		self.lsps2_client.as_ref().map(|s| (s.lsp_node_id, s.lsp_address.clone()))
 	}
 
 	pub(crate) fn lsps2_channel_needs_manual_broadcast(
@@ -763,283 +739,6 @@ where
 		}
 	}
 
-	pub(crate) async fn lsps2_receive_to_jit_channel(
-		&self, amount_msat: u64, description: &Bolt11InvoiceDescription, expiry_secs: u32,
-		max_total_lsp_fee_limit_msat: Option<u64>, payment_hash: Option<PaymentHash>,
-	) -> Result<Bolt11Invoice, Error> {
-		let fee_response = self.lsps2_request_opening_fee_params().await?;
-
-		let (min_total_fee_msat, min_opening_params) = fee_response
-			.opening_fee_params_menu
-			.into_iter()
-			.filter_map(|params| {
-				if amount_msat < params.min_payment_size_msat
-					|| amount_msat > params.max_payment_size_msat
-				{
-					log_debug!(self.logger,
-						"Skipping LSP-offered JIT parameters as the payment of {}msat doesn't meet LSP limits (min: {}msat, max: {}msat)",
-						amount_msat,
-						params.min_payment_size_msat,
-						params.max_payment_size_msat
-					);
-					None
-				} else {
-					compute_opening_fee(amount_msat, params.min_fee_msat, params.proportional as u64)
-						.map(|fee| (fee, params))
-				}
-			})
-			.min_by_key(|p| p.0)
-			.ok_or_else(|| {
-				log_error!(self.logger, "Failed to handle response from liquidity service",);
-				Error::LiquidityRequestFailed
-			})?;
-
-		if let Some(max_total_lsp_fee_limit_msat) = max_total_lsp_fee_limit_msat {
-			if min_total_fee_msat > max_total_lsp_fee_limit_msat {
-				log_error!(self.logger,
-					"Failed to request inbound JIT channel as LSP's requested total opening fee of {}msat exceeds our fee limit of {}msat",
-					min_total_fee_msat, max_total_lsp_fee_limit_msat
-				);
-				return Err(Error::LiquidityFeeTooHigh);
-			}
-		}
-
-		log_debug!(
-			self.logger,
-			"Choosing cheapest liquidity offer, will pay {}msat in total LSP fees",
-			min_total_fee_msat
-		);
-
-		let buy_response =
-			self.lsps2_send_buy_request(Some(amount_msat), min_opening_params).await?;
-		let lsps2_parameters = LSPS2Parameters {
-			max_total_opening_fee_msat: Some(min_total_fee_msat),
-			max_proportional_opening_fee_ppm_msat: None,
-		};
-		let invoice = self.lsps2_create_jit_invoice(
-			buy_response,
-			Some(amount_msat),
-			description,
-			expiry_secs,
-			payment_hash,
-			lsps2_parameters,
-		)?;
-
-		log_info!(self.logger, "JIT-channel invoice created: {}", invoice);
-		Ok(invoice)
-	}
-
-	pub(crate) async fn lsps2_receive_variable_amount_to_jit_channel(
-		&self, description: &Bolt11InvoiceDescription, expiry_secs: u32,
-		max_proportional_lsp_fee_limit_ppm_msat: Option<u64>, payment_hash: Option<PaymentHash>,
-	) -> Result<Bolt11Invoice, Error> {
-		let fee_response = self.lsps2_request_opening_fee_params().await?;
-
-		let (min_prop_fee_ppm_msat, min_opening_params) = fee_response
-			.opening_fee_params_menu
-			.into_iter()
-			.map(|params| (params.proportional as u64, params))
-			.min_by_key(|p| p.0)
-			.ok_or_else(|| {
-				log_error!(self.logger, "Failed to handle response from liquidity service",);
-				Error::LiquidityRequestFailed
-			})?;
-
-		if let Some(max_proportional_lsp_fee_limit_ppm_msat) =
-			max_proportional_lsp_fee_limit_ppm_msat
-		{
-			if min_prop_fee_ppm_msat > max_proportional_lsp_fee_limit_ppm_msat {
-				log_error!(self.logger,
-					"Failed to request inbound JIT channel as LSP's requested proportional opening fee of {} ppm msat exceeds our fee limit of {} ppm msat",
-					min_prop_fee_ppm_msat,
-					max_proportional_lsp_fee_limit_ppm_msat
-				);
-				return Err(Error::LiquidityFeeTooHigh);
-			}
-		}
-
-		log_debug!(
-			self.logger,
-			"Choosing cheapest liquidity offer, will pay {}ppm msat in proportional LSP fees",
-			min_prop_fee_ppm_msat
-		);
-
-		let buy_response = self.lsps2_send_buy_request(None, min_opening_params).await?;
-		let lsps2_parameters = LSPS2Parameters {
-			max_total_opening_fee_msat: None,
-			max_proportional_opening_fee_ppm_msat: Some(min_prop_fee_ppm_msat),
-		};
-		let invoice = self.lsps2_create_jit_invoice(
-			buy_response,
-			None,
-			description,
-			expiry_secs,
-			payment_hash,
-			lsps2_parameters,
-		)?;
-
-		log_info!(self.logger, "JIT-channel invoice created: {}", invoice);
-		Ok(invoice)
-	}
-
-	async fn lsps2_request_opening_fee_params(&self) -> Result<LSPS2FeeResponse, Error> {
-		let lsps2_client = self.lsps2_client.as_ref().ok_or(Error::LiquiditySourceUnavailable)?;
-
-		let client_handler = self.liquidity_manager.lsps2_client_handler().ok_or_else(|| {
-			log_error!(self.logger, "Liquidity client was not configured.",);
-			Error::LiquiditySourceUnavailable
-		})?;
-
-		let (fee_request_sender, fee_request_receiver) = oneshot::channel();
-		{
-			let mut pending_fee_requests_lock =
-				lsps2_client.pending_fee_requests.lock().expect("lock");
-			let request_id = client_handler
-				.request_opening_params(lsps2_client.lsp_node_id, lsps2_client.token.clone());
-			pending_fee_requests_lock.insert(request_id, fee_request_sender);
-		}
-
-		tokio::time::timeout(
-			Duration::from_secs(LIQUIDITY_REQUEST_TIMEOUT_SECS),
-			fee_request_receiver,
-		)
-		.await
-		.map_err(|e| {
-			log_error!(self.logger, "Liquidity request timed out: {}", e);
-			Error::LiquidityRequestFailed
-		})?
-		.map_err(|e| {
-			log_error!(self.logger, "Failed to handle response from liquidity service: {}", e);
-			Error::LiquidityRequestFailed
-		})
-	}
-
-	async fn lsps2_send_buy_request(
-		&self, amount_msat: Option<u64>, opening_fee_params: LSPS2OpeningFeeParams,
-	) -> Result<LSPS2BuyResponse, Error> {
-		let lsps2_client = self.lsps2_client.as_ref().ok_or(Error::LiquiditySourceUnavailable)?;
-
-		let client_handler = self.liquidity_manager.lsps2_client_handler().ok_or_else(|| {
-			log_error!(self.logger, "Liquidity client was not configured.",);
-			Error::LiquiditySourceUnavailable
-		})?;
-
-		let (buy_request_sender, buy_request_receiver) = oneshot::channel();
-		{
-			let mut pending_buy_requests_lock =
-				lsps2_client.pending_buy_requests.lock().expect("lock");
-			let request_id = client_handler
-				.select_opening_params(lsps2_client.lsp_node_id, amount_msat, opening_fee_params)
-				.map_err(|e| {
-					log_error!(
-						self.logger,
-						"Failed to send buy request to liquidity service: {:?}",
-						e
-					);
-					Error::LiquidityRequestFailed
-				})?;
-			pending_buy_requests_lock.insert(request_id, buy_request_sender);
-		}
-
-		let buy_response = tokio::time::timeout(
-			Duration::from_secs(LIQUIDITY_REQUEST_TIMEOUT_SECS),
-			buy_request_receiver,
-		)
-		.await
-		.map_err(|e| {
-			log_error!(self.logger, "Liquidity request timed out: {}", e);
-			Error::LiquidityRequestFailed
-		})?
-		.map_err(|e| {
-			log_error!(self.logger, "Failed to handle response from liquidity service: {:?}", e);
-			Error::LiquidityRequestFailed
-		})?;
-
-		Ok(buy_response)
-	}
-
-	fn lsps2_create_jit_invoice(
-		&self, buy_response: LSPS2BuyResponse, amount_msat: Option<u64>,
-		description: &Bolt11InvoiceDescription, expiry_secs: u32,
-		payment_hash: Option<PaymentHash>, lsps2_parameters: LSPS2Parameters,
-	) -> Result<Bolt11Invoice, Error> {
-		let lsps2_client = self.lsps2_client.as_ref().ok_or(Error::LiquiditySourceUnavailable)?;
-
-		// LSPS2 requires min_final_cltv_expiry_delta to be at least 2 more than usual.
-		let min_final_cltv_expiry_delta = MIN_FINAL_CLTV_EXPIRY_DELTA + 2;
-		let encoded_payment_metadata =
-			PaymentMetadata { lsps2_parameters: Some(lsps2_parameters) }.encode();
-		let (payment_hash, payment_secret, payment_metadata) = match payment_hash {
-			Some(payment_hash) => {
-				let (payment_secret, payment_metadata) = self
-					.channel_manager
-					.create_inbound_payment_for_hash(
-						payment_hash,
-						None,
-						expiry_secs,
-						Some(min_final_cltv_expiry_delta),
-						Some(encoded_payment_metadata),
-					)
-					.map_err(|e| {
-						log_error!(self.logger, "Failed to register inbound payment: {:?}", e);
-						Error::InvoiceCreationFailed
-					})?;
-				(payment_hash, payment_secret, payment_metadata)
-			},
-			None => self
-				.channel_manager
-				.create_inbound_payment(
-					None,
-					expiry_secs,
-					Some(min_final_cltv_expiry_delta),
-					Some(encoded_payment_metadata),
-				)
-				.map_err(|e| {
-					log_error!(self.logger, "Failed to register inbound payment: {:?}", e);
-					Error::InvoiceCreationFailed
-				})?,
-		};
-
-		let route_hint = RouteHint(vec![RouteHintHop {
-			src_node_id: lsps2_client.lsp_node_id,
-			short_channel_id: buy_response.intercept_scid,
-			fees: RoutingFees { base_msat: 0, proportional_millionths: 0 },
-			cltv_expiry_delta: buy_response.cltv_expiry_delta as u16,
-			htlc_minimum_msat: None,
-			htlc_maximum_msat: None,
-		}]);
-
-		let currency = self.config.network.into();
-		let mut invoice_builder = InvoiceBuilder::new(currency)
-			.invoice_description(description.clone())
-			.payment_hash(payment_hash)
-			.payment_secret(payment_secret)
-			.current_timestamp()
-			.min_final_cltv_expiry_delta(min_final_cltv_expiry_delta.into())
-			.expiry_time(Duration::from_secs(expiry_secs.into()))
-			.private_route(route_hint);
-
-		if let Some(amount_msat) = amount_msat {
-			invoice_builder = invoice_builder.amount_milli_satoshis(amount_msat).basic_mpp();
-		}
-
-		let invoice = if let Some(payment_metadata) = payment_metadata {
-			invoice_builder.payment_metadata(payment_metadata).build_signed(|hash| {
-				Secp256k1::new()
-					.sign_ecdsa_recoverable(hash, &self.keys_manager.get_node_secret_key())
-			})
-		} else {
-			invoice_builder.build_signed(|hash| {
-				Secp256k1::new()
-					.sign_ecdsa_recoverable(hash, &self.keys_manager.get_node_secret_key())
-			})
-		};
-		invoice.map_err(|e| {
-			log_error!(self.logger, "Failed to build and sign invoice: {}", e);
-			Error::InvoiceCreationFailed
-		})
-	}
-
 	pub(crate) async fn handle_channel_ready(
 		&self, user_channel_id: u128, channel_id: &ChannelId, counterparty_node_id: &PublicKey,
 	) {
@@ -1109,15 +808,4 @@ where
 			}
 		}
 	}
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct LSPS2FeeResponse {
-	opening_fee_params_menu: Vec<LSPS2OpeningFeeParams>,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct LSPS2BuyResponse {
-	intercept_scid: u64,
-	cltv_expiry_delta: u32,
 }
